% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{action_disallow}
\alias{action_disallow}
\title{Disallow an action}
\usage{
action_disallow(pkgs, db, envir = parent.frame())
}
\description{
Disallow an action
}
\seealso{
Other actions: 
\code{\link{default_actions}()},
\code{\link{last_rejected}()},
\code{\link{last_rejected_permit}()}
}
\concept{actions}
