% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{default_actions}
\alias{default_actions}
\title{Default actions}
\usage{
default_actions()
}
\value{
A data.frame with three columns: functions, arguments and actions.
}
\description{
Decide what to do when some functions are called.
}
\examples{
df <- default_actions()
}
