% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{last_rejected}
\alias{last_rejected}
\title{Retrieve the last package rejection, imposed by a criteria}
\usage{
last_rejected()
}
\description{
Retrieve the last package rejection, imposed by a criteria
}
\seealso{
Other actions: 
\code{\link{action_disallow}()},
\code{\link{default_actions}()},
\code{\link{last_rejected_permit}()}
}
\concept{actions}
