% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{last_rejected_permit}
\alias{last_rejected_permit}
\title{Permit the last package rejection}
\usage{
last_rejected_permit(quiet = FALSE)
}
\description{
Permit the last package rejection
}
\seealso{
Other actions: 
\code{\link{action_disallow}()},
\code{\link{default_actions}()},
\code{\link{last_rejected}()}
}
\concept{actions}
