% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\title{val.criterion Options}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{val.criterion} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "val.criterion")

options::opt(x, default, env = "val.criterion")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{repos}{\describe{
Metrics repositories to use implicitly for package filters. Will be
accessed using \code{\link[=available.packages]{available.packages()}}, so repository paths should follow
similar conventions.\item{default: }{\preformatted{character(0L)}}
\item{option: }{val.criterion.repos}
\item{envvar: }{R_VAL_CRITERION_REPOS (evaluated if possible, raw string otherwise)}
}}

\item{exceptions}{\describe{
Packages that should be treated as exceptions. Even if they would otherwise
be filtered out using preset filters, they would still be discoverable and
installable without issue.\item{default: }{\preformatted{character(0L)}}
\item{option: }{val.criterion.exceptions}
\item{envvar: }{R_VAL_CRITERION_EXCEPTIONS (evaluated if possible, raw string otherwise)}
}}

\item{quiet}{\describe{
Wether filtering operations should be reported as they're being queried.
Passed to the \code{quiet} parameter of individual functions where it's used.\item{default: }{\preformatted{FALSE}}
\item{option: }{val.criterion.quiet}
\item{envvar: }{R_VAL_CRITERION_QUIET (evaluated if possible, raw string otherwise)}
}}

\item{actions}{\describe{
A set of calls that should prompt actions. See \code{\link{default_actions}} for an
example of the expected format.\item{default: }{\preformatted{default_actions()}}
\item{option: }{val.criterion.actions}
\item{envvar: }{R_VAL_CRITERION_ACTIONS (evaluated if possible, raw string otherwise)}
}}

}
}

\seealso{
options getOption Sys.setenv Sys.getenv
}
