% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options_params}
\alias{options_params}
\title{Options As Parameters}
\arguments{
\item{quiet}{Wether filtering operations should be reported as they're being queried.
Passed to the \code{quiet} parameter of individual functions where it's used. (Defaults to \code{FALSE}, overwritable using option 'val.criterion.quiet' or environment variable 'R_VAL_CRITERION_QUIET')}

\item{monitor}{A set of calls that should trigger actions. See \code{\link{default_action_calls}}
for (Defaults to \code{default_monitored_calls}, overwritable using option 'val.criterion.monitor' or environment variable 'R_VAL_CRITERION_MONITOR')}

\item{repos}{Metrics repositories to use implicitly for package filters. Will be
accessed using \code{\link[=available.packages]{available.packages()}}, so repository paths should follow
similar conventions. (Defaults to \code{character(0L)}, overwritable using option 'val.criterion.repos' or environment variable 'R_VAL_CRITERION_REPOS')}

\item{install.action}{Describe What action should be taken when a filtered package is accessed.
\code{install} actions are taken when known package-installing functions are
used and hook in to \code{available.packages()} to retrieve a package index. (Defaults to \code{warn_filter_exception}, overwritable using option 'val.criterion.install.action' or environment variable 'R_VAL_CRITERION_INSTALL_ACTION')}

\item{exceptions}{Packages that should be treated as exceptions. Even if they would otherwise
be filtered out using preset filters, they would still be discoverable and
installable without issue. (Defaults to \code{character(0L)}, overwritable using option 'val.criterion.exceptions' or environment variable 'R_VAL_CRITERION_EXCEPTIONS')}
}
\description{
Options As Parameters
}
\keyword{internal}
