% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.val.criterion::filter}
\alias{plot.val.criterion::filter}
\title{Plot the decision tree of a criterion filter}
\usage{
\method{plot}{`val.criterion::filter`}(x, db = x(as = "data.frame"), ...)
}
\arguments{
\item{x}{A \code{\link{filter}} object, created using \code{\link{package_filter}} for use with
\code{\link[=available.packages]{available.packages()}}.}

\item{db}{Optionally, a \code{\link[=available.packages]{available.packages()}} database to use when
plotting package counts at each stage of the decision process. By default,
will use the \code{\link[=available.packages]{available.packages()}} output as though the filter had not
been applied.}
}
\value{
either \code{integer} node ID if a node was added, or a
\code{character} label if the current expression cannot be added as a node.
}
\description{
Plot the decision tree of a criterion filter
}
