% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cooldown.R
\name{cooldown}
\alias{cooldown}
\title{Filter by date}
\usage{
cooldown(accepted = Sys.Date() - 2 * 7, ...)
}
\arguments{
\item{accepted}{A date of some time in the past until which published
packages are accepted.}

\item{...}{Other arguments passed to \code{\link{package_filter}}.}
}
\value{
A \code{\link{package_filter}}
}
\description{
Implement a cooldown filter so that only packages that are older than a given
threshold are installed.
}
\details{
This helps to prevent installing packages recently published with a bug or an
infiltration. For the same reason it prevents installing updates and patches
of recently fixed packages.
}
\examples{
# entire available packages set
ap_complete <- available.packages()
nrow(ap_complete)

# available packages with cooldown filter applied
ap <- available.packages(fields = "Published", filters = cooldown())
ap <- subset(as.data.frame(ap), as.logical(Include))
nrow(ap)

}
