% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{default_actions}
\alias{default_actions}
\title{A default set of package tools to impose actions upon}
\usage{
default_actions()
}
\description{
A default set of package tools to impose actions upon
}
\seealso{
Other actions: 
\code{\link{action_disallow}()},
\code{\link{last_rejected}()},
\code{\link{last_rejected_permit}()}
}
\concept{actions}
