% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options_params}
\alias{options_params}
\title{Options As Parameters}
\arguments{
\item{quiet}{Wether filtering operations should be reported as they're being queried.
Passed to the \code{quiet} parameter of individual functions where it's used. (Defaults to \code{FALSE}, overwritable using option 'val.criterion.quiet' or environment variable 'R_VAL_CRITERION_QUIET')}

\item{repos}{Metrics repositories to use implicitly for package filters. Will be
accessed using \code{\link[=available.packages]{available.packages()}}, so repository paths should follow
similar conventions. (Defaults to \code{character(0L)}, overwritable using option 'val.criterion.repos' or environment variable 'R_VAL_CRITERION_REPOS')}

\item{exceptions}{Packages that should be treated as exceptions. Even if they would otherwise
be filtered out using preset filters, they would still be discoverable and
installable without issue. (Defaults to \code{character(0L)}, overwritable using option 'val.criterion.exceptions' or environment variable 'R_VAL_CRITERION_EXCEPTIONS')}

\item{actions}{A set of calls that should prompt actions. See \code{\link{default_actions}} for an
example of the expected format. (Defaults to \code{default_actions()}, overwritable using option 'val.criterion.actions' or environment variable 'R_VAL_CRITERION_ACTIONS')}
}
\description{
Options As Parameters
}
\keyword{internal}
