% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{val.criterion-package}
\alias{val.criterion}
\alias{val.criterion-package}
\title{val.criterion: Tools for filtering repositories based on expressed criteria.}
\description{
Provides high-level tools for specifying filtering criteria to use when pulling packages and mechanisms of hooking into R's internal tools for interfacing with package repositories to impose constraints on the available packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmar.github.io/val.criterion}
  \item \url{https://github.com/pharmar/val.criterion}
  \item \url{https://pharmar.org/}
  \item Report bugs at \url{https://github.com/pharmaR/val.criterion/issues}
}

}
\author{
\strong{Maintainer}: Doug Kelkhoff \email{doug.kelkhoff@gmail.com}

Other contributors:
\itemize{
  \item R Validation Hub \email{r-val-hub-maintainers@lists.r-consortium.org} (https://www.pharmar.org) [copyright holder]
}

}
\keyword{internal}
